move TrackableParser packets/structs into respective folders
