+++ source/include/protocol/cs_TrackableParserPackets.h
@@ -6,6 +6,8 @@
  */
 
 #pragma once
+#include <cstdint>
+#include <third/cuckoo/CuckooFilter.h>
 
 // ------------------ Command values -----------------
 
@@ -47,4 +49,40 @@ struct __attribute__((__packed__)) trackable_parser_cmd_commit_filter_changes_re
 
 struct __attribute__((__packed__)) trackable_parser_cmd_get_filer_summaries_ret_t {
 
+};
+
+// ------------------ Filter format ------------------
+
+/**
+ * Determines whether the filter has mac or advertisement data as input.
+ */
+enum class FilterInputType : uint8_t {
+	MacAddress = 0,
+	AdData = 1,
+};
+
+/**
+ * Data associated to a parsing filter that is persisted to flash.
+ */
+struct __attribute__((__packed__)) TrackingFilterMetaData {
+public:
+	// sync info
+	uint8_t protocol; // determines implementation type of filter
+	uint16_t version;
+
+	uint8_t profileId; // devices passing the filter are assigned this profile id
+	FilterInputType inputType;
+
+	// sync state
+	union{
+		uint8_t flags;
+		struct {
+			uint8_t isActive : 1;
+		} bits;
+	};
+};
+
+struct __attribute__((__packed__)) TrackingFilterData {
+	TrackingFilterMetaData metadata;
+	CuckooFilter filter;
 };
